[consolidate-create-constants] Moves constants, gets rid of poplar constants
