+++ src/forest/jasmine/sogp_gps.py
@@ -13,17 +13,14 @@ Optimization work: Eli Jones
 """
 
 import logging
-import math
 
 import numba
 import numpy as np
 
+from forest.constants import SECONDS_PER_DAY_TIMES_PI, SECONDS_PER_WEEK_TIMES_PI
+
 
 logger = logging.getLogger(__name__)
-
-
-SECONDS_PER_DAY_TIMES_PI = 86_400 * math.pi
-SECONDS_PER_WEEK_TIMES_PI = 604_800 * math.pi
 
 
 @numba.njit(cache=True, fastmath=True)
